:recycle: refactor : AWS S3 코드 분리
